Make ObjectRefGuard::nuketype_mut() private.
I definitely don't like the idea right now of being able to change an object's
type. Feel like it would be nice to be able to rely on the assumption that it
won't change, and then perhaps in the future we can add the type representation
to `ObjectBox` so we can check it without locking the object. Not that I really
think that would be beneficial, but there might be some use in the future for
such a thing.